Simulate a DECAL perturbation experiment

It simulates a random DECAL experiment generating a random UMI count matrix with cells randomly assigned to different clones and introducing to each clone perturbations to random genes.

sim_experiment(
  depth,
  ratio,
  lfc = 0,
  nclones = 10L,
  min_n = 2L,
  max_n = 20L,
  theta = 100L
)

sim_experiment_from_data(
  reference,
  lfc = 0,
  nclones = 10L,
  min_n = 2L,
  max_n = 20L,
  theta = 100L,
  ngenes = NULL
)

Arguments

depth
cells total UMI count

ratio
ratio of genes UMI over total UMI count

lfc
vector indicating the perturbations to be generated for each clone

nclones
number of clones to be simulated

min_n
minimum clone size

max_n
maximum clone size

theta
negative binomial dispersion parameter

reference
UMI count matrix used as base for the simulation

ngenes
number of genes to be simulated. When NULL it replicates the gene ratio found in reference, otherwise it simulates ngenes with ratio ranging in a logarithm scale from lowest to highest observed ratio observed.

Value

a list containing the following fields:

perturbations, a table indicating the perturbations expected_lfc for a gene in a clone.

clone, a list of cells assigned to each clone.

count, a random UMI count matrix.